+++ controller/apisensorloggercontroller.php
@@ -352,12 +352,14 @@ class ApiSensorLoggerController extends ApiController
      */
     protected function insertDeviceType($deviceType)
     {
-        $sql = 'INSERT INTO `*PREFIX*sensorlogger_device_types` (`user_id`,`device_type_name`) VALUES(?,?)';
-        $stmt = $this->db->prepare($sql);
-        $stmt->bindParam(1, $this->userSession->getUser()->getUID());
-        $stmt->bindParam(2, $deviceType);
-        if ($stmt->execute()) {
-            return (int)$this->db->lastInsertId();
+        $userId = $this->userSession->getUser()->getUID();
+        $queryBuilder = $this->db->getQueryBuilder();
+        $queryBuilder->insert('sensorlogger_device_types')
+            ->setValue('device_type_name', $queryBuilder->createNamedParameter($deviceType))
+            ->setValue('user_id', $queryBuilder->createNamedParameter($userId));
+
+        if ($queryBuilder->execute()) {
+            return (int)$queryBuilder->getLastInsertId();
         }
         return false;
     }
@@ -368,12 +370,14 @@ class ApiSensorLoggerController extends ApiController
      */
     protected function insertDeviceGroup($string)
     {
-        $sql = 'INSERT INTO `*PREFIX*sensorlogger_device_groups` (`user_id`,`device_group_name`) VALUES(?,?)';
-        $stmt = $this->db->prepare($sql);
-        $stmt->bindParam(1, $this->userSession->getUser()->getUID());
-        $stmt->bindParam(2, $string);
-        if ($stmt->execute()) {
-            return (int)$this->db->lastInsertId();
+        $userId = $this->userSession->getUser()->getUID();
+        $queryBuilder = $this->db->getQueryBuilder();
+        $queryBuilder->insert('sensorlogger_device_groups')
+            ->setValue('device_group_name', $queryBuilder->createNamedParameter($string))
+            ->setValue('user_id', $queryBuilder->createNamedParameter($userId));
+
+        if ($queryBuilder->execute()) {
+            return (int)$queryBuilder->getLastInsertId();
         }
         return false;
     }
@@ -384,8 +388,8 @@ class ApiSensorLoggerController extends ApiController
      */
     protected function insertDataTypes($array)
     {
-        $sql = 'INSERT INTO `*PREFIX*sensorlogger_data_types` (`user_id`,`description`,`type`,`short`) VALUES(?,?,?,?)';
-        $stmt = $this->db->prepare($sql);
+//        $sql = 'INSERT INTO `*PREFIX*sensorlogger_data_types` (`user_id`,`description`,`type`,`short`) VALUES(?,?,?,?)';
+//        $stmt = $this->db->prepare($sql);
 
         if (isset($array['description'])) {
             $description = $array['description'] ?: '';
@@ -405,12 +409,16 @@ class ApiSensorLoggerController extends ApiController
             $unit = '';
         }
 
-        $stmt->bindParam(1, $this->userSession->getUser()->getUID());
-        $stmt->bindParam(2, $description);
-        $stmt->bindParam(3, $type);
-        $stmt->bindParam(4, $unit);
-        if ($stmt->execute()) {
-            return (int)$this->db->lastInsertId();
+        $userId = $this->userSession->getUser()->getUID();
+        $queryBuilder = $this->db->getQueryBuilder();
+        $queryBuilder->insert('sensorlogger_data_types')
+            ->setValue('user_id', $queryBuilder->createNamedParameter($userId))
+            ->setValue('description', $queryBuilder->createNamedParameter($description))
+            ->setValue('type', $queryBuilder->createNamedParameter($type))
+            ->setValue('short', $queryBuilder->createNamedParameter($unit));
+
+        if ($queryBuilder->execute()) {
+            return (int)$queryBuilder->getLastInsertId();
         }
         return false;
     }
@@ -461,12 +469,21 @@ class ApiSensorLoggerController extends ApiController
      */
     protected function insertDeviceDataTypes($deviceId, $dataTypeId)
     {
-        $sql = 'INSERT INTO `*PREFIX*sensorlogger_device_data_types` (`user_id`,`device_id`,`data_type_id`) VALUES(?,?,?)';
-        $stmt = $this->db->prepare($sql);
-        $stmt->bindParam(1, $this->userSession->getUser()->getUID());
-        $stmt->bindParam(2, $deviceId);
-        $stmt->bindParam(3, $dataTypeId);
-        $stmt->execute();
+//        $sql = 'INSERT INTO `*PREFIX*sensorlogger_device_data_types` (`user_id`,`device_id`,`data_type_id`) VALUES(?,?,?)';
+//        $stmt = $this->db->prepare($sql);
+//        $stmt->bindParam(1, $this->userSession->getUser()->getUID());
+//        $stmt->bindParam(2, $deviceId);
+//        $stmt->bindParam(3, $dataTypeId);
+//        $stmt->execute();
+
+        $userId = $this->userSession->getUser()->getUID();
+        $queryBuilder = $this->db->getQueryBuilder();
+        $queryBuilder->insert('sensorlogger_device_data_types')
+            ->setValue('device_id', $queryBuilder->createNamedParameter($deviceId))
+            ->setValue('data_type_id', $queryBuilder->createNamedParameter($dataTypeId))
+            ->setValue('user_id', $queryBuilder->createNamedParameter($userId));
+
+        $queryBuilder->execute();
     }
 
     /**
@@ -520,9 +537,6 @@ class ApiSensorLoggerController extends ApiController
      */
     protected function insertDevice($array)
     {
-        $sql = 'INSERT INTO `*PREFIX*sensorlogger_devices` (`uuid`,`name`,`type_id`,`user_id`) VALUES(?,?,?,?)';
-        $stmt = $this->db->prepare($sql);
-
         if (isset($array['deviceId'])) {
             if (!isset($array['deviceName'])) {
                 $array['deviceName'] = 'Default device';
@@ -532,13 +546,16 @@ class ApiSensorLoggerController extends ApiController
                 $array['deviceTypeId'] = 0;
             }
 
-            $stmt->bindParam(1, $array['deviceId']);
-            $stmt->bindParam(2, $array['deviceName']);
-            $stmt->bindParam(3, $array['deviceTypeId']);
-            $stmt->bindParam(4, $this->userSession->getUser()->getUID());
+            $userId = $this->userSession->getUser()->getUID();
+            $queryBuilder = $this->db->getQueryBuilder();
+            $queryBuilder->insert('sensorlogger_devices')
+                ->setValue('uuid', $queryBuilder->createNamedParameter($array['deviceId']))
+                ->setValue('name', $queryBuilder->createNamedParameter($array['deviceName']))
+                ->setValue('type_id', $queryBuilder->createNamedParameter($array['deviceTypeId']))
+                ->setValue('user_id', $queryBuilder->createNamedParameter($userId));
 
-            if ($stmt->execute()) {
-                return (int)$this->db->lastInsertId();
+            if ($queryBuilder->execute()) {
+                return (int)$queryBuilder->getLastInsertId();
             }
         } else {
             return 'Missing device ID';